Define fixed 24.576 MHz clock in device tree.
This is for the external 24.576MHz clock on the beaglebone black
baseboard. This is needed for the PCM512x codec driver. This is a hack as
the clock is currently controlled from davinci-evm.c via a GPIO and therefore
the clock will not be present if that driver is not running.