ASoc: pcm512x: Add a clock-type specification to device tree bindings
The pcm512x and related devices have an on chip PLL that can be used to
generate an audio master clock from virtually any frequency reference
clock. This commit adds the necessary device tree bindings to allow the user
to select this mode. For backwards compatibility the default is to assume
that the clock is a multiple of the sampling frequency. The funtionality for
the PLL mode will be added in future commits.

Signed-off-by: Howard Mitchell <hm@hmbedded.co.uk>